Pre-Launch DD Event + Exception Propagation Fix

+++ src/Tgstation.Server.Host.Watchdog/Watchdog.cs
@@ -177,6 +177,10 @@ namespace Tgstation.Server.Host.Watchdog
 
 				var watchdogVersion = executingAssembly.GetName().Version?.Semver().ToString();
 
+				var serializerOptions = new JsonSerializerOptions
+				{
+					WriteIndented = true,
+				};
 				while (!cancellationToken.IsCancellationRequested)
 				{
 					if (!File.Exists(assemblyPath))
@@ -214,10 +218,7 @@ namespace Tgstation.Server.Host.Watchdog
 								bootstrapperSettingsFile,
 								JsonSerializer.Serialize(
 									bootstrapSettings,
-									new JsonSerializerOptions
-									{
-										WriteIndented = true,
-									}),
+									serializerOptions),
 								cancellationToken);
 						}
 					}
@@ -258,6 +259,7 @@ namespace Tgstation.Server.Host.Watchdog
 							process.StartInfo.UseShellExecute = false; // runs in the same console
 
 							var killedHostProcess = false;
+							var createdShutdownFile = false;
 							try
 							{
 								Task? processTask = null;
@@ -279,6 +281,7 @@ namespace Tgstation.Server.Host.Watchdog
 									{
 										logger.LogInformation("Cancellation requested! Writing shutdown lock file...");
 										File.WriteAllBytes(updateDirectory, Array.Empty<byte>());
+										createdShutdownFile = true;
 									}
 									else
 										logger.LogWarning("Cancellation requested while update directory exists!");
@@ -330,15 +333,16 @@ namespace Tgstation.Server.Host.Watchdog
 									logger.LogWarning(ex2, "Error killing host process!");
 								}
 
-								try
-								{
-									if (File.Exists(updateDirectory))
-										File.Delete(updateDirectory);
-								}
-								catch (Exception ex2)
-								{
-									logger.LogWarning(ex2, "Error deleting comms file!");
-								}
+								if (createdShutdownFile)
+									try
+									{
+										if (File.Exists(updateDirectory))
+											File.Delete(updateDirectory);
+									}
+									catch (Exception ex2)
+									{
+										logger.LogWarning(ex2, "Error deleting comms file!");
+									}
 
 								logger.LogInformation("Host exited!");
 							}
